requirements.txt
APScheduler==3.10.4
lxml==5.2.2
Pillow==10.4.0
psutil==6.0.0
pystray==0.19.5
Requests==2.32.3
urllib3==2.2.2
